+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/Rules/SubEntities.hs
@@ -2,6 +2,7 @@
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE GeneralisedNewtypeDeriving #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE OverloadedStrings #-}
@@ -105,19 +106,19 @@ instance
 
 data SubEntitiesPredFailure era
   = SubCertsFailure (PredicateFailure (EraRule "SUBCERTS" era))
-  | SubMissingAccountsInWithdrawals Withdrawals
-  | SubMissingOriginalAccountsInWithdrawals Withdrawals
-  | SubMissingAccountsInDirectDeposits DirectDeposits
-  | SubWrongNetworkInWithdrawals
+  | SubWithdrawalAddressesWithWrongNetwork
       -- | Expected network id
       Network
-      -- | Withdrawal accounts with wrong network id
+      -- | Withdrawal account addresses with wrong network id
       (NonEmptySet AccountAddress)
-  | SubWrongNetworkInDirectDeposits
+  | SubWithdrawalAccountsMissing Withdrawals
+  | SubWithdrawalAccountsMissingPreBatch Withdrawals
+  | SubDirectDepositAddressesWithWrongNetwork
       -- | Expected network id
       Network
-      -- | Direct-deposit accounts with wrong network id
+      -- | Direct-deposit account addresses with wrong network id
       (NonEmptySet AccountAddress)
+  | SubDirectDepositAccountsMissing DirectDeposits
   | SubBalancesOutsideAccountBalanceIntervals
       (NonEmptyMap AccountAddress (Coin, AccountBalanceInterval era))
   | SubMissingAccountsInAccountBalanceIntervals
@@ -147,11 +148,11 @@ instance
   encCBOR =
     encode . \case
       SubCertsFailure x -> Sum (SubCertsFailure @era) 0 !> To x
-      SubMissingAccountsInWithdrawals x -> Sum (SubMissingAccountsInWithdrawals @era) 1 !> To x
-      SubMissingOriginalAccountsInWithdrawals x -> Sum (SubMissingOriginalAccountsInWithdrawals @era) 2 !> To x
-      SubMissingAccountsInDirectDeposits x -> Sum (SubMissingAccountsInDirectDeposits @era) 3 !> To x
-      SubWrongNetworkInWithdrawals expected wrongs -> Sum (SubWrongNetworkInWithdrawals @era) 4 !> To expected !> To wrongs
-      SubWrongNetworkInDirectDeposits expected wrongs -> Sum (SubWrongNetworkInDirectDeposits @era) 5 !> To expected !> To wrongs
+      SubWithdrawalAddressesWithWrongNetwork expected wrongs -> Sum (SubWithdrawalAddressesWithWrongNetwork @era) 1 !> To expected !> To wrongs
+      SubWithdrawalAccountsMissing x -> Sum (SubWithdrawalAccountsMissing @era) 2 !> To x
+      SubWithdrawalAccountsMissingPreBatch x -> Sum (SubWithdrawalAccountsMissingPreBatch @era) 3 !> To x
+      SubDirectDepositAddressesWithWrongNetwork expected wrongs -> Sum (SubDirectDepositAddressesWithWrongNetwork @era) 4 !> To expected !> To wrongs
+      SubDirectDepositAccountsMissing x -> Sum (SubDirectDepositAccountsMissing @era) 5 !> To x
       SubBalancesOutsideAccountBalanceIntervals x -> Sum (SubBalancesOutsideAccountBalanceIntervals @era) 6 !> To x
       SubMissingAccountsInAccountBalanceIntervals x -> Sum (SubMissingAccountsInAccountBalanceIntervals @era) 7 !> To x
       SubWrongNetworkInAccountBalanceIntervals expected wrongs -> Sum (SubWrongNetworkInAccountBalanceIntervals @era) 8 !> To expected !> To wrongs
@@ -164,11 +165,11 @@ instance
   where
   decCBOR = decode . Summands "SubEntitiesPredFailure" $ \case
     0 -> SumD SubCertsFailure <! From
-    1 -> SumD SubMissingAccountsInWithdrawals <! From
-    2 -> SumD SubMissingOriginalAccountsInWithdrawals <! From
-    3 -> SumD SubMissingAccountsInDirectDeposits <! From
-    4 -> SumD SubWrongNetworkInWithdrawals <! From <! From
-    5 -> SumD SubWrongNetworkInDirectDeposits <! From <! From
+    1 -> SumD SubWithdrawalAddressesWithWrongNetwork <! From <! From
+    2 -> SumD SubWithdrawalAccountsMissing <! From
+    3 -> SumD SubWithdrawalAccountsMissingPreBatch <! From
+    4 -> SumD SubDirectDepositAddressesWithWrongNetwork <! From <! From
+    5 -> SumD SubDirectDepositAccountsMissing <! From
     6 -> SumD SubBalancesOutsideAccountBalanceIntervals <! From
     7 -> SumD SubMissingAccountsInAccountBalanceIntervals <! From
     8 -> SumD SubWrongNetworkInAccountBalanceIntervals <! From <! From
@@ -284,7 +285,7 @@ validateMissingOriginalAccountsInWithdrawals ::
 validateMissingOriginalAccountsInWithdrawals wdrls originalAccounts =
   failureOnJust
     (withdrawalsMissingAccounts wdrls originalAccounts)
-    SubMissingOriginalAccountsInWithdrawals
+    SubWithdrawalAccountsMissingPreBatch
 
 validateMissingAccountsInWithdrawals ::
   EraAccounts era =>
@@ -294,7 +295,7 @@ validateMissingAccountsInWithdrawals ::
 validateMissingAccountsInWithdrawals wdrls accounts =
   failureOnJust
     (withdrawalsMissingAccounts wdrls accounts)
-    SubMissingAccountsInWithdrawals
+    SubWithdrawalAccountsMissing
 
 conwayToDijkstraSubEntitiesPredFailure ::
   forall era. Conway.ConwayLedgerPredFailure era -> SubEntitiesPredFailure era
@@ -314,16 +315,16 @@ conwayToDijkstraSubEntitiesPredFailure = \case
 entitiesToSubEntitiesPredFailure ::
   EntitiesPredFailure era -> SubEntitiesPredFailure era
 entitiesToSubEntitiesPredFailure = \case
-  WrongNetworkInWithdrawals net addrs -> SubWrongNetworkInWithdrawals net addrs
-  WrongNetworkInDirectDeposits net addrs -> SubWrongNetworkInDirectDeposits net addrs
+  WithdrawalAddressesWithWrongNetwork net addrs -> SubWithdrawalAddressesWithWrongNetwork net addrs
+  DirectDepositAddressesWithWrongNetwork net addrs -> SubDirectDepositAddressesWithWrongNetwork net addrs
+  DirectDepositAccountsMissing dds -> SubDirectDepositAccountsMissing dds
+  CertsFailure _ -> impossible "CertsFailure"
+  WithdrawalAccountsMissing _ -> impossible "WithdrawalAccountsMissing"
+  WithdrawalAmountsInexactInLegacyMode _ -> impossible "WithdrawalAmountsInexactInLegacyMode"
+  WithdrawalAmountsExceedingOriginalBalance _ -> impossible "WithdrawalAmountsExceedingOriginalBalance"
   WrongNetworkInAccountBalanceIntervals net addrs -> SubWrongNetworkInAccountBalanceIntervals net addrs
   MissingAccountsInAccountBalanceIntervals x -> SubMissingAccountsInAccountBalanceIntervals x
   BalancesOutsideAccountBalanceIntervals x -> SubBalancesOutsideAccountBalanceIntervals x
-  CertsFailure _ -> impossible "CertsFailure"
-  MissingAccountsInWithdrawals _ -> impossible "MissingAccountsInWithdrawals"
-  IncompleteWithdrawals _ -> impossible "IncompleteWithdrawals"
-  ExceededBalancesInWithdrawals _ -> impossible "ExceededBalancesInWithdrawals"
-  MissingAccountsInDirectDeposits dds -> SubMissingAccountsInDirectDeposits dds
   WrongNetworkInStartingAccountBalanceIntervals _ _ ->
     impossible "WrongNetworkInStartingAccountBalanceIntervals"
   MissingAccountsInStartingAccountBalanceIntervals _ ->